refactor: #118 - hooks 경로 수정
