+++ src/components/PostListItem/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useDeletePost } from './useDeletePost';
