+++ src/pages/SearchPage/hooks/index.ts
@@ -1,0 +1,2 @@
+export { default as usePostSort } from './usePostSort';
+export { default as useUsersort } from './useUserSort';
